refactor(proxy) - rearrange internal logic
