+++ internal/proxy/server.go
@@ -1,4 +1,4 @@
-package server
+package proxy
 
 import (
 	"net/http"
